feat(drawer): allow children to be a render function
Adds support for a render function children in `Drawer`. This allows content inside the drawer to be rendered on-demand and destroyed when the drawer is closed. Notably, such behavior cannot be accomplished by conditioning on `isOpen` because `isOpen` would destroy the contents too quickly and we want to wait until the drawer is fully closed.

See storybook example for a more descriptive demonstration.